requirements.txt
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
certifi
cffi
charset-normalizer
click
colorama
cryptography
dataclasses-json
Deprecated
deprecation
distro
dnspython
email-validator
exceptiongroup
fastapi
fastapi-cli
frozenlist
greenlet
h11
httpcore
httptools
httpx
Jinja2
MarkupSafe
marshmallow
mdurl
multidict
openai
packaging
pdfminer.six
pikepdf
Pillow
pluggy
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pymongo
python-dotenv
python-multipart
PyYAML
reportlab
requests
rich
shellingham
sniffio
SQLAlchemy
starlette
tenacity
tqdm
typer
typing-inspect
typing_extensions
urllib3
uvicorn
watchfiles
websockets
wrapt
yarl
boto3
jwt